+++ src/App.tsx
@@ -1,6 +1,5 @@
 import RouteIndex from "./components/pages/RouteIndex";
-// import RouteEditor from "./components/pages/RouteEditor";
-import RouteViewer from "./components/pages/RouteViewer";
+import RoutePage from "./components/pages/Routes/Route";
 import SignIn from "./components/pages/SignIn";
 import SignUp from "./components/pages/SignUp";
 import Top from "./components/pages/Top";
@@ -26,7 +25,7 @@ function App() {
           <Route exact path={pagePaths.SIGN_IN} children={<SignIn />} />
           <Route exact path={pagePaths.SIGN_UP} children={<SignUp />} />
           <Route exact path={pagePaths.ROUTE_NEW} children={<RouteNew />} />
-          <Route path={pagePaths.ROUTE_EDITOR} children={<RouteViewer />} />
+          <Route path={pagePaths.ROUTE_EDITOR} children={<RoutePage />} />
           <Route
             exact
             path={pagePaths.PASSWORD_RESET}
